requirements.txt
Django>=5.0
djangorestframework>=3.15
scikit-learn>=1.4
xgboost>=2.0
joblib>=1.3
numpy>=1.26
pandas>=2.2
gunicorn>=21.2
